blog_management_system/requirements.txt
flask==2.3.3
flask-cors==4.0.0
flask-sqlalchemy==3.0.5
werkzeug==2.3.7
pyjwt==2.8.0
python-dotenv==1.0.0
gunicorn==21.2.0